Conflicts:
	public/cquiz/index.html
	public/js/createQuizApp.js
	public/js/quizApp.js
	scripts/quiz.js

+++ scripts/quiz.js
@@ -1,34 +1,50 @@
 var token;
-function login() {
+
+function login(toggleLogin) {
   var type = "redirect";
   var url = "http://www.quizalize.com/quiz#/";
   Zzish.init("2d14d1984a2e3293bd13aab34c85e2ea");
 
+
   token = localStorage.getItem("zzishtoken");
   email = localStorage.getItem("emailAddress");
   if (token==null && email==null) {
-      Zzish.login(type,url,function(err,message) {
-          $("#LoginButton").html("Logout");
-          console.log("Logged in with status and message",err,message);
-          if (!err) {
-              loginUser(message);
-          }
-          else {
-              console.log("Error",err);
-          }
-      });
-   }
-   else if (email!=null) {
+      Zzish.login(type,url);
+  }
+  else if (email!=null) {
           localStorage.clear();
           $("#LoginButton").html("Login with Zzish");
           location.href="/quiz/";
-   }
-   else {
+  }
+  else {
+    if (!!toggleLogin) {
       Zzish.logout(token,function(err,message) {
           localStorage.clear();
           $("#LoginButton").html("Login with Zzish");
           location.href="/quiz/";
       });
+    }
+    else {
+      location.href=url;
+    }
+  }
+}
+
+function logout() {
+  var url = "http://www.quizalize.com/quiz#/";
+  Zzish.init("2d14d1984a2e3293bd13aab34c85e2ea");
+
+  token = localStorage.getItem("zzishtoken");
+  if (token!=null) {
+      Zzish.logout(token,function(err,message) {
+          localStorage.clear();
+          $("#LoginButton").html("Login with Zzish");
+          location.href="/quiz/";
+      });
+  }
+  else {
+    localStorage.clear();
+    location.href="/quiz/";
   }
 }
